refactor: move remaining flat modules into subfolders
Relocate api, errors, permissions, policies, types, and colour into
domain folders with direct imports throughout. Split published error
classes into separate files (no max-classes-per-file ignore). Update
tsdown and deno.json export paths; npm subpaths unchanged.
