+++ __tests__/errors.test.ts
@@ -5,7 +5,7 @@ import {
   RestrictedPermissionError,
   TokenCreationError,
   TokenDeletionError,
-} from "#src/errors.ts";
+} from "#src/errors/index.ts";
 
 describe("CloudflareApiError", () => {
   test.concurrent("sets _tag to CloudflareApiError", () => {
