+++ src/automation/discovery.ts
@@ -5,7 +5,11 @@
  */
 
 import type { OutputFormat } from "#src/cli/args.ts";
-import type { Account, PermissionGroup, ServiceGroup } from "#src/types.ts";
+import type {
+  Account,
+  PermissionGroup,
+  ServiceGroup,
+} from "#src/types/index.ts";
 
 interface ScopeListEntry {
   access: {
