+++ src/cli/help.ts
@@ -10,7 +10,7 @@ import {
   readAutomationFile,
   SKILL_REFERENCE_FILES,
 } from "#src/automation/paths.ts";
-import colour from "#src/colour.ts";
+import colour from "#src/terminal/colour.ts";
 
 const NAME = "create-cf-token";
 const VERSION = process.env.npm_package_version ?? "0.0.0";
