+++ src/errors/cloudflare-api-error.ts
@@ -1,0 +1,10 @@
+import { CloudflareApiErrorBase } from "#src/errors/bases.ts";
+
+export class CloudflareApiError extends CloudflareApiErrorBase {
+  constructor(args: { path: string; messages: string[] }) {
+    super({
+      ...args,
+      message: `CF API error (${args.path}): ${args.messages.join(", ")}`,
+    });
+  }
+}
