+++ src/errors/restricted-permission-error.ts
@@ -1,0 +1,10 @@
+import { RestrictedPermissionErrorBase } from "#src/errors/bases.ts";
+
+export class RestrictedPermissionError extends RestrictedPermissionErrorBase {
+  constructor(args: { permissionName: string; errorText: string }) {
+    super({
+      ...args,
+      message: `Restricted permission: ${args.permissionName}`,
+    });
+  }
+}
