+++ src/errors/token-creation-error.ts
@@ -1,0 +1,10 @@
+import { TokenCreationErrorBase } from "#src/errors/bases.ts";
+
+export class TokenCreationError extends TokenCreationErrorBase {
+  constructor(args: { errorText: string }) {
+    super({
+      ...args,
+      message: `Token creation failed: ${args.errorText}`,
+    });
+  }
+}
