+++ src/errors/token-deletion-error.ts
@@ -1,0 +1,10 @@
+import { TokenDeletionErrorBase } from "#src/errors/bases.ts";
+
+export class TokenDeletionError extends TokenDeletionErrorBase {
+  constructor(args: { errorText: string }) {
+    super({
+      ...args,
+      message: `Token deletion failed: ${args.errorText}`,
+    });
+  }
+}
