+++ src/policies/build.ts
@@ -1,11 +1,15 @@
 /**
- * @module policies
+ * @module policies/build
  *
  * Cloudflare API token policy construction from permission selections.
  */
 
-import { isPermissionExcluded } from "#src/permissions.ts";
-import type { Account, PermissionGroup, TokenPolicy } from "#src/types.ts";
+import { isPermissionExcluded } from "#src/permissions/group.ts";
+import type {
+  Account,
+  PermissionGroup,
+  TokenPolicy,
+} from "#src/types/index.ts";
 
 export function buildPolicies(
   perms: PermissionGroup[],
