+++ tsdown.config.ts
@@ -12,16 +12,16 @@ export default defineConfig({
   },
   dts: true,
   entry: {
-    api: "src/api.ts",
+    api: "src/api/client.ts",
     cli: "src/cli.ts",
     create: "src/automation/create.ts",
-    errors: "src/errors.ts",
+    errors: "src/errors/index.ts",
     index: "src/index.ts",
-    permissions: "src/permissions.ts",
-    policies: "src/policies.ts",
+    permissions: "src/permissions/group.ts",
+    policies: "src/policies/build.ts",
     "scope-spec": "src/automation/scope-spec.ts",
     spec: "src/automation/spec.ts",
-    types: "src/types.ts",
+    types: "src/types/index.ts",
   },
   format: "esm",
   outDir: "dist",
